+++ Engine/Vec2.h
@@ -4,14 +4,14 @@ class Vec2
 {
 public:
 	Vec2() = default;
-	Vec2(int in_x, int in_y);
+	Vec2(float in_x, float in_y);
 	Vec2 operator+(const Vec2& rhs) const;
 	Vec2 operator-(const Vec2& rhs) const;
 	Vec2& operator+=(const Vec2& rhs);
 	Vec2& operator-=(const Vec2& rhs);
-	Vec2 operator*( const int rhs) const;
-	Vec2 operator/( const int rhs) const;
+	Vec2 operator*( const float rhs) const;
+	Vec2 operator/( const float rhs) const;
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 };
